Renamed `enum` package to `mode`.
Fixes #1
