+++ kexasol/ExaConnectionOptions.kt
@@ -1,6 +1,6 @@
 package com.badoo.kexasol
 
-import com.badoo.kexasol.enum.ExaEncryptionMode
+import com.badoo.kexasol.mode.ExaEncryptionMode
 import com.badoo.kexasol.net.ExaNodeAddress
 import com.badoo.kexasol.stream.ExaStreamWorker
