+++ kexasol/stream/ExaStreamHttpThread.kt
@@ -1,7 +1,7 @@
 package com.badoo.kexasol.stream
 
 import com.badoo.kexasol.ExaConnectionOptions
-import com.badoo.kexasol.enum.ExaEncryptionMode
+import com.badoo.kexasol.mode.ExaEncryptionMode
 import com.badoo.kexasol.net.ExaNodeAddress
 import com.badoo.kexasol.net.ExaSelfSignedSslSocketFactory
 import okio.*
